% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_simulation.R
\name{si_simulation}
\alias{si_simulation}
\title{Susceptible(S) & Infected(I) model simulation}
\usage{
si_simulation(g, spreaders, prob, layout = layout_nicely(g), plot_iter = F)
}
\arguments{
\item{g}{Graph object.}

\item{spreaders}{List of initial spreaders (graph vertices).}

\item{prob}{Probability of infecting neighbour.}

\item{layout}{Custom layout, default is layout_nicely.}

\item{plot_iter}{Boolean argument, default FALSE. If TRUE plot every iteration of spreading.}
}
\value{
Number of iterations to infect all nodes.
}
\description{
Initially, all nodes are susceptible status except for a set of
             r infected nodes selected as source spreaders. At each
             time step, infected node tries to infect one of its neighbors
             with given probability.
}
\examples{
library(igraphdata)

data(karate)
spreaders <- voterank(karate, 2)
si_simulation(karate, spreaders, prob = 0.5)
}
