% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wvoterank.R
\name{wvoterank}
\alias{wvoterank}
\title{WVoteRank}
\usage{
wvoterank(
  g,
  r,
  display = F,
  display_layout = layout_nicely(g),
  display_label = T
)
}
\arguments{
\item{g}{Weighted graph object.}

\item{r}{Number of spreaders.}

\item{display}{Whether to display graph, default FALSE.}

\item{display_layout}{Custom layout of graph, dafault layout_nicely.}

\item{display_label}{Whether to display vertex.label, default TRUE.}
}
\value{
Vector of \code{r} vertices identified as spreaders.
}
\description{
A voting approach to uncover multiple influential spreaders
             on weighted networks
}
\examples{
library(igraphdata)

data(karate)
wvoterank(karate, 2, weight = E(karate)$weight)

}
\references{
Csardi G, Nepusz T: The igraph software package for complex
            network research, InterJournal, Complex Systems 1695. 2006.
            \url{https://igraph.org}

Hong-liang, S. et al.: A voting approach to uncover multiple
            influential spreaders on weighted networks. Physica A:
            Statistical Mechanics and its Applications Volume 519, 2019,
            Pages 303-312. \url{https://doi.org/10.1016/j.physa.2018.12.001}
}
