% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voterank.R
\name{voterank}
\alias{voterank}
\title{VoteRank}
\usage{
voterank(
  g,
  r,
  display = F,
  display_layout = layout_nicely(g),
  display_label = T
)
}
\arguments{
\item{g}{Graph object.}

\item{r}{Number of spreaders.}

\item{display}{Whether to display graph, default FALSE.}

\item{display_layout}{Custom layout of graph, dafault layout_nicely.}

\item{display_label}{Whether to display vertex.label, default TRUE.}
}
\value{
Vector of \code{r} vertices identified as spreaders.
}
\description{
VoteRank algorithm for identifying a set of influential
             spreaders in complex networks.
}
\examples{
library(igraphdata)

data(karate)
voterank(karate, 2)

}
\references{
Csardi G, Nepusz T: The igraph software package for complex
            network research, InterJournal, Complex Systems 1695. 2006.
            \url{https://igraph.org}

Zhang, JX., Chen, DB., Dong, Q. et al.: Identifying a set of
            influential spreaders in complex networks. Sci Rep 6, 27823
            (2016). \url{https://doi.org/10.1038/srep27823}
            (\url{https://www.nature.com/articles/srep27823.pdf})
}
