% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voterank.R
\name{voterank}
\alias{voterank}
\title{VoteRank}
\usage{
voterank(g, r)
}
\arguments{
\item{g}{Graph object.}

\item{r}{Number of spreaders.}
}
\value{
Vector of \code{r} vertices identified as spreaders.
}
\description{
VoteRank algorithm for identifying a set of influential spreaders in complex networks.
}
\examples{
library(igraphdata)

data(karate)
voterank(karate, 2)

}
\references{
Csardi G, Nepusz T: The igraph software package for complex network research, InterJournal, Complex Systems 1695. 2006. \url{https://igraph.org}

Panfeng Liu, Longjie Li, Shiyu Fang, Yukai Yao: Identifying influential nodes in social networks: A voting approach, Chaos, Solitons & Fractals 152. 2021. \url{https://doi.org/10.1016/j.chaos.2021.111309} (\url{https://www.sciencedirect.com/science/article/pii/S0960077921006639})

Zhang, JX., Chen, DB., Dong, Q. et al.: Identifying a set of influential spreaders in complex networks. Sci Rep 6, 27823 (2016). \url{https://doi.org/10.1038/srep27823} (\url{https://www.nature.com/articles/srep27823.pdf})
}
